# Conflicts:
#	source/Game.vcxproj.filters
#	source/components/player/comp_player_model.h

+++ source/components/player/states/attack_states/VerticalLauncherActionState.h
@@ -17,7 +17,7 @@ class VerticalLauncherActionState : public GroundedActionState {
 	float interruptibleTime;
 
 	CHandle hitboxHandle;
-	int damage = 0;
+	float damage = 0;
 
 public:
 	VerticalLauncherActionState(TCompPlayerModel* player, CHandle hitbox);
